refactor(comments): extract channel formatting into posthog.comment
Move the channel-agnostic content conversion (TipTap rich_content <-> markdown,
Slack mrkdwn/blocks, email HTML) out of the conversations product into a new core
posthog.comment package so core discussions and other products can reuse it without
importing product internals. Pure move; behavior unchanged.

Generated-By: PostHog Code
Task-Id: 9ac92446-4916-4c82-80ba-7f554be8f479
